Finish healing and adding characters
Healing is now fully finished (except for overheal but that will be for later). I have now also added all the characters I want to use with Castorice and have added some light cones to the sim. Will work on adding healing to the characters that have it and making Sunday work with Castorice.